mmap -> multiple pmap
